+++ public/CHANGELOG.md
@@ -10,6 +10,7 @@ Although this tries to adhere to "Semantic Versioning", this is a website, not a
 ## [1.6.4] 2017-04-17 - 2017-04-19
 ### Changed
 - `[Backend]` Site classes now have their own files, as per good PHP practice.
+- An alert is now shown to the user if they try to update anything on the site but their session has expired.
 
 ### Fixed
 - MAL Sync properly works with volumes marked with TBD/TBA/NA without preceding v.